fix(hivemind): error handling, SQL type casts, and RAG pipeline fixes
- Add proper HTTP status codes (400/401/409) for auth validation errors
- Fix usage summary SQL: cast SUM() results to bigint to prevent NUMERIC/INT8 mismatch panic
- Fix Qdrant connection: use gRPC port 6334 with proper API key auth
- Recreate knowledge collection with correct 768 dimensions for nomic-embed-text
- Add PDF document upload support with reportlab-compatible extraction
- Add knowledge search handler with meeting context joins
- Fix member registration error handling for missing invites
- Add comprehensive integration test script (56 tests, all passing)

+++ src/repos/usage.rs
@@ -47,10 +47,10 @@ impl UsageRepo {
         let rows = sqlx::query(
             r#"
             SELECT tu.user_id,
-                   SUM(tu.input_tokens) as total_input_tokens,
-                   SUM(tu.output_tokens) as total_output_tokens,
-                   SUM(tu.cost_cents) as total_cost_cents,
-                   COUNT(DISTINCT tu.session_id) as session_count,
+                   SUM(tu.input_tokens)::bigint as total_input_tokens,
+                   SUM(tu.output_tokens)::bigint as total_output_tokens,
+                   SUM(tu.cost_cents)::bigint as total_cost_cents,
+                   COUNT(DISTINCT tu.session_id)::bigint as session_count,
                    MAX(tu.recorded_at) as last_active_at
             FROM token_usage tu
             WHERE tu.company_id = $1
